+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.0.4 | 
-- Consolidate unzip decorator to handle .grib2.gz and .grib2 files.
+- Consolidate unzip decorator to handle both `.grib2.gz` and `.grib2` files.
 - Add example notebooks.
+- Add other data sources from MRMS: precipitation rate, precipitation flag, and 1h, 24h and 72h accumulated precipitation. 
 
 
 # ⭐ 0.0.3 | 2024-03-25